Add files via upload
I'm not sure if the meta files are needed or not, but having all of these together should work.
I modified the testmap scene by adding a character, player, and status script. In this scene, a HUD should appear and a health bar is used as a pseudo player. You can move the "player" around with WASD, lower and raise the health with IO, and do the same for the mana with KL.