Sam Merging some new damage scripts
Updated enemies movements, implemented 2 damage types. now damage is sent as (object).sendMessage.("PhysicalDamage", float damage);
currently this can be sent to the Player script and the EnemyCharacterStats script as well. substitute "MagicalDamage" as the first sent variable to sent it as magic dmg instead. Keep in mind defenses will apply on sent damage so you don't need to apply it before sending.